requirements.txt
django<1.10
django-debug-toolbar<1.5
django-multiupload<0.6